users_service/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
packaging==25.0
psycopg2==2.9.11
pyasn1==0.6.1
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==4.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
wrapt==2.0.1
zipp==3.23.0
